Update README.md
Updating BooksOnTheFly app with a new idea. Create a simple web server program which will request the contents of a web page and convert it over to a new format. Similar to Readable or one of the web browser options, which strips a page until it only contains the article.